+++ data/schema.graphql
@@ -7367,6 +7367,7 @@ input MyCollectionCreateArtworkInput {
   height: String
   medium: String!
   metric: String
+  provenance: String
   title: String
   width: String
 }
@@ -7410,6 +7411,7 @@ input MyCollectionUpdateArtworkInput {
   height: String
   medium: String
   metric: String
+  provenance: String
   title: String
   width: String
 }
